+++ src/main/java/com/ddabong/ddabongdotchiBE/domain/user/service/UserService.java
@@ -9,13 +9,17 @@ import com.ddabong.ddabongdotchiBE.domain.s3.util.S3Service;
 import com.ddabong.ddabongdotchiBE.domain.user.dto.request.PasswordUpdateRequest;
 import com.ddabong.ddabongdotchiBE.domain.user.dto.request.UserJoinRequest;
 import com.ddabong.ddabongdotchiBE.domain.user.dto.request.UserUpdateRequest;
+import com.ddabong.ddabongdotchiBE.domain.user.dto.response.ReissueResponse;
 import com.ddabong.ddabongdotchiBE.domain.user.dto.response.UserJoinResponse;
 import com.ddabong.ddabongdotchiBE.domain.user.dto.response.UserUpdateResponse;
 import com.ddabong.ddabongdotchiBE.domain.user.entity.User;
+import com.ddabong.ddabongdotchiBE.domain.user.jwt.util.JwtUtil;
 import com.ddabong.ddabongdotchiBE.domain.user.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @RequiredArgsConstructor
 @Transactional
 @Service
@@ -24,6 +28,7 @@ public class UserService {
 	private final UserRepository userRepository;
 	private final PasswordEncoder passwordEncoder;
 	private final S3Service s3Service;
+	private final JwtUtil jwtUtil;
 
 	public UserJoinResponse join(UserJoinRequest request, MultipartFile file) {
 		String imageUrl = s3Service.uploadImage(file);
@@ -31,6 +36,11 @@ public class UserService {
 		user.setImageUrl(imageUrl);
 		userRepository.save(user);
 		return UserJoinResponse.from(user);
+	}
+
+	public ReissueResponse reissueToken(String refreshToken) {
+		log.info("[*] Generate new token pair with " + refreshToken);
+		return ReissueResponse.from(jwtUtil.reissueToken(refreshToken));
 	}
 
 	public void deactivate(User user) {
